ビルド落ちるバグ修正
